guard main-execution
The Developing Modules guide mentions that the main function should be guarded
(as is the common idiom for Python scripts).  Most of the modules were not
written that way.  This patch updates most of them to comply with this policy.

See here: http://docs.ansible.com/ansible/developing_modules.html#common-module-boilerplate